+++ server/src/api/sites/batchImportEvents.ts
@@ -2,9 +2,9 @@ import { FastifyReply, FastifyRequest } from "fastify";
 import { z } from "zod";
 import { clickhouse } from "../../db/clickhouse/clickhouse.js";
 import { updateImportProgress, completeImport, getImportById } from "../../services/import/importStatusManager.js";
-import { UmamiEvent, UmamiImportMapper } from "../../services/import/mappings/umami.js";
-import { SimpleAnalyticsEvent, SimpleAnalyticsImportMapper } from "../../services/import/mappings/simpleAnalytics.js";
-import { PlausibleEvent, PlausibleImportMapper } from "../../services/import/mappings/plausible.js";
+import { UmamiEvent, UmamiImportMapper } from "../../services/import/mappers/umami.js";
+import { SimpleAnalyticsEvent, SimpleAnalyticsImportMapper } from "../../services/import/mappers/simpleAnalytics.js";
+import { PlausibleEvent, PlausibleImportMapper } from "../../services/import/mappers/plausible.js";
 import { importQuotaManager } from "../../services/import/importQuotaManager.js";
 import { db } from "../../db/postgres/postgres.js";
 import { organization, sites } from "../../db/postgres/schema.js";
