+++ server/src/services/import/mappers/simpleAnalytics.ts
